comment changes, new intents, method return change
Changed getGarageInfo() method so that instead of immediately returning the dictionary, it returns the two lists created in the method. 

Changed the Garage intent so that now users can ask for a specific garage's capacity, and added a method two zip the two lists from getGarageInfo() into a dictionary for this method. 

Added a method for the AllGarages intent which is called when a user wants to see the capacities for all of the garages, which is originally what the script did.